adjust integration tests
