+++ das_client/app/integration_test/mocks/mock_authenticated_scope.dart
@@ -1,13 +1,14 @@
 // implement the mock for authenticated_scope.dart
 
 import 'package:app/di/di.dart';
-import 'package:app/di/scopes/authenticated_scope.dart';
+import 'package:app/pages/journey/view_model/warn_app_view_model.dart';
 import 'package:app/provider/ru_feature_provider.dart';
 import 'package:formation/component.dart';
 import 'package:logging/logging.dart';
 
 import 'mock_formation_repository.dart';
 import 'mock_ru_feature_provider.dart';
+import 'mock_warn_app_view_model.dart';
 
 final _log = Logger('MockAuthenticatedScope');
 
@@ -26,10 +27,11 @@ class MockAuthenticatedScope extends AuthenticatedScope {
 
     getIt.registerAuthProvider();
     getIt.registerSferaAuthProvider();
-    getIt.registerMqttAuthProvider();
-    getIt.registerMqttService();
     getIt.registerHttpClient();
-    getIt.registerSferaRemoteRepo();
+    getIt.registerMqttAuthProvider();
+    await getIt.registerMqttService();
+    await getIt.registerSferaRemoteRepo();
+    getIt.registerAppExpirationViewModel();
     if (e2e) {
       getIt.registerSettingsRepository();
       getIt.registerRuFeatureProvider();
@@ -38,7 +40,13 @@ class MockAuthenticatedScope extends AuthenticatedScope {
       _registerMockRuFeaturesProvider();
       _registerMockFormationRepository();
     }
-    getIt.registerAppExpirationViewModel();
+
+    getIt.registerJourneyNavigationViewModel();
+    getIt.registerJourneySelectionViewModel();
+    getIt.registerJourneyViewModel();
+    getIt.registerNotificationPriorityViewModel();
+    _registerMockWarnAppViewModel();
+    getIt.registerLocalRegulationHtmlGenerator();
 
     return getIt.allReady();
   }
@@ -49,5 +57,18 @@ class MockAuthenticatedScope extends AuthenticatedScope {
 
   void _registerMockFormationRepository() {
     getIt.registerSingleton<FormationRepository>(MockFormationRepository());
+  }
+
+  void _registerMockWarnAppViewModel() {
+    getIt.registerSingleton<WarnAppViewModel>(
+      MockWarnAppViewModel(
+        flavor: DI.get(),
+        sferaRepo: DI.get(),
+        warnappRepo: DI.get(),
+        ruFeatureProvider: DI.get(),
+        notificationViewModel: DI.get(),
+      ),
+      dispose: (vm) => vm.dispose(),
+    );
   }
 }
